% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeysOutlier.R
\name{ANOVA2_BW}
\alias{ANOVA2_BW}
\title{This method assumes two independent variables and a single dependent variable.}
\usage{
ANOVA2_BW(x, y, z)
}
\arguments{
\item{x}{An independent variable}

\item{y}{Another independent variable}

\item{z}{the dependent variable}
}
\value{
The resulting ANOVA from \code{x} and \code{y} and \code{z}
}
\description{
This method assumes two independent variables and a single dependent variable.
}
\examples{
remotes::install_github("brittn27/RepPack27")
library(RepPack27)
data(RTDATA)
x <- RTDATA$difficulty
y <- RTDATA$depth
z <- RTDATA$RT
Data<- RTDATA
ANOVA2_BW(x,y,z)

}
