% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeysOutlier.R
\name{LeysOutlier}
\alias{LeysOutlier}
\title{remove data for RT outside the median absolute deviation method.
This method assumes two independent variables and a single dependent variable.}
\usage{
LeysOutlier(x, y, z)
}
\description{
param x An independent variable
param y Another independent variable
param z the dependent variable
export
}
\examples{
# example code
x <- Data$IV1
y <- Data$IV2
z <- Data$DV


}
