% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepPack27-package.R
\docType{package}
\name{RepPack27-package}
\alias{RepPack27}
\alias{RepPack27-package}
\title{RepPack27: Noah's Package for Github 712}
\description{
This package was created with the intention of cleaning data from psychopy to make future analysis easier.

It was also created for the course Github 712 to demonstrate
an example for how to create one relevant to my present research.
}
\references{
\url{https://github.com/brittn27/RepPack27}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/brittn27/RepPack27}
  \item Report bugs at \url{https://github.com/brittn27/RepPack27/issues}
}

}
\author{
\strong{Maintainer}: Noah Britt \email{brittn@mcmaster.ca} (\href{https://orcid.org/0000-0001-5020-4528}{ORCID})


Noah Britt, Department of Psychology, Neuroscience, and Behaviour, McMaster University \email{brittn@mcmaster.ca}
}
\keyword{internal}
