% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RepPack27-package.R
\docType{data}
\name{RTDATA}
\alias{RTDATA}
\title{Reaction Time for a Random Participant in my RT Study using Discrimination in Depth.}
\format{
A data frame with 672 trials across 9 important variables post-cleaning:
\describe{
\item{Picture}{Whichever condition was presented on a given trial}
\item{difficulty}{This notes the orientation of the discriminated lines}
\item{depth}{Whether the lines were presented in near/far space}
\item{side}{left or right side presentation}
\item{elevation}{Where was this line presented relative to the reference line}
\item{correctAns}{The expected response if discriminated properly}
\item{key_resp.corr}{If the participant responded correctly}
\item{key_resp.rt}{Reaction Time}
\item{participant}{Random Participant ID allocated}

}
}
\source{
SunLAB McMaster University via Hongjin Sun (PNB)
}
\usage{
data(RTDATA)
}
\description{
A dataset containing RT data across each condition. Orientation of the discriminated lines and depth information.
}
\examples{
 data(RTDATA)
}
\keyword{datasets}
