% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeysOutlier.R
\name{LeysOutlier}
\alias{LeysOutlier}
\title{remove data for RT outside the median absolute deviation method.
This method assumes two independent variables and a single dependent variable.}
\usage{
LeysOutlier(x, y, z)
}
\arguments{
\item{x}{An independent variable}

\item{y}{Another independent variable}

\item{z}{the dependent variable}
}
\value{
The outlier method of \code{x} and \code{y} and \code{z}
}
\description{
remove data for RT outside the median absolute deviation method.
This method assumes two independent variables and a single dependent variable.
}
\examples{
# example code
x <- Data$IV1
y <- Data$IV2
z <- Data$DV
OriginalData <- Data
LeysOutlier(x,y,z)
NewData # This should get you your new dataset


}
